requirements.txt
##pip install -r requirements.txt
google-generativeai
Pillow